examples/agent/6-appsource/agent/requirements.txt
google_adk==1.4.1
fastapi==0.115.13
uvicorn==0.34.0
pydantic==2.11.7
